Added HTML elements for signin/login page and added styles with css
So far, I have added the static HTML elements on the page. I have added some styling to the elements based on the figma file. Till this commit, I left off on writing the media query for screens larger than 700px. Right now, I still have a bug where, as the size of the screen increases, the margins of the login wrapper begin to touch the size of the page. Moreover, the paddings of the 'Sign In' button is a bit off.